requirements.txt
annotated-types==0.7.0
anyio==4.8.0
argcomplete==1.10.0
astroid==3.3.8
asttokens==3.0.0
Authlib==1.4.0
bcrypt==4.2.1
beautifulsoup4==4.8.0
black==24.10.0
boto3==1.36.4
botocore==1.36.4
Brotli==1.1.0
cachetools==5.5.2
certifi==2024.12.14
cffi==1.17.1
cfgv==3.4.0
chardet==3.0.4
charset-normalizer==3.4.1
click==8.1.8
contourpy==1.3.1
cryptography==44.0.0
cssselect2==0.8.0
cycler==0.12.1
decorator==5.1.1
dill==0.3.9
distlib==0.3.9
distro==1.9.0
dnspython==2.7.0
docstring_parser==0.16
docx2txt==0.8
email_validator==2.2.0
et_xmlfile==2.0.0
executing==2.1.0
extract-msg==0.23.1
fastapi==0.115.6
fastapi-cli==0.0.7
filelock==3.16.1
fonttools==4.56.0
fsspec==2025.2.0
google-api-core==2.24.1
google-auth==2.38.0
google-cloud-aiplatform==1.71.1
google-cloud-bigquery==3.30.0
google-cloud-core==2.4.2
google-cloud-resource-manager==1.14.1
google-cloud-storage==2.19.0
google-crc32c==1.6.0
google-genai==1.5.0
google-resumable-media==2.7.2
googleapis-common-protos==1.69.1
greenlet==3.1.1
groq==0.18.0
grpc-google-iam-v1==0.14.1
grpcio==1.70.0
grpcio-status==1.70.0
gTTS==2.5.4
h11==0.14.0
html2text==2024.2.26
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
identify==2.6.5
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
IMAPClient==2.1.0
iniconfig==2.0.0
ipython==8.31.0
isort==5.13.2
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.5
jiter==0.8.2
jmespath==1.0.1
jose==1.0.0
kiwisolver==1.4.8
llvmlite==0.44.0
lxml==5.3.0
mariadb==1.1.11
markdown-it-py==3.0.0
markdown2==2.5.3
MarkupSafe==3.0.2
matplotlib==3.10.0
matplotlib-inline==0.1.7
mccabe==0.7.0
mdurl==0.1.2
more-itertools==10.6.0
moviepy==2.1.2
mpmath==1.3.0
mypy-extensions==1.0.0
networkx==3.4.2
nodeenv==1.9.1
numba==0.61.0
numpy==2.1.3
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
olefile==0.46
openai==1.60.2
openpyxl==3.1.5
orjson==3.10.14
packaging==24.2
pandas==2.2.3
parso==0.8.4
passlib==1.7.4
pathspec==0.12.1
pexpect==4.9.0
phonenumbers==8.13.52
pillow==10.4.0
pincode==1.0.3
platformdirs==4.3.6
pluggy==1.5.0
pre_commit==4.0.1
proglog==0.1.10
prompt_toolkit==3.0.48
proto-plus==1.26.0
protobuf==5.29.3
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pycryptodome==3.21.0
pydantic==2.10.5
pydantic-extra-types==2.10.1
pydantic-settings==2.7.1
pydantic_core==2.27.2
pydub==0.25.1
pydyf==0.11.0
Pygments==2.19.1
pylint==3.3.3
pyparsing==3.2.1
PyPDF2==3.0.1
pyphen==0.17.2
pytest==8.3.4
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.20
python-pptx==0.6.18
python-stdnum==1.20
pyttsx3==2.98
pytz==2024.2
PyYAML==6.0.2
regex==2024.11.6
reportlab==4.3.1
requests==2.32.3
rich==13.9.4
rich-toolkit==0.12.0
rsa==4.9
s3transfer==0.11.1
setuptools==75.8.0
shapely==2.0.7
shellingham==1.5.4
sniffio==1.3.1
sortedcontainers==2.4.0
sounddevice==0.5.1
soupsieve==2.6
SpeechRecognition==3.8.1
SQLAlchemy==2.0.37
stack-data==0.6.3
starlette==0.41.3
sympy==1.13.1
tiktoken==0.8.0
tinycss2==1.4.0
tinyhtml5==2.0.0
tomlkit==0.13.2
torch==2.6.0
tqdm==4.67.1
traitlets==5.14.3
triton==3.2.0
typer==0.15.1
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==1.5.1
ujson==5.10.0
urllib3==2.3.0
uvicorn==0.34.0
uvloop==0.21.0
validators==0.34.0
virtualenv==20.28.1
watchfiles==1.0.4
wcwidth==0.2.13
weasyprint==64.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==14.1
Werkzeug==3.1.3
wheel==0.45.1
xlrd==1.2.0
XlsxWriter==3.2.2
zopfli==0.2.3.post1
python-magic